feat: extract mod configs page to separate component
